refactor: introduce ScoreManagerService
